fix(reports): move submitReportSchema out of use-server module
Next 16.3 rejects non-async-function exports from a use-server file; the
zod submitReportSchema object export broke server-action evaluation on the
listing page POST (A use server file can only export async functions,
found object). Move the schema into lib/reports/constants (pure module)
and have the action + unit test import it from there.

Also folds in the pre-existing uncommitted removal of the now-unused
SELLER_REPORT_REASONS constant (report-dialog uses REPORT_REASONS since 8b82724).
