+++ web/lib/reports/constants.ts
@@ -8,6 +8,8 @@
  * here (`export *`).
  */
 
+import { z } from "zod"
+
 export const REPORT_REASONS = [
   "spam",
   "fraud",
@@ -18,15 +20,6 @@ export const REPORT_REASONS = [
 ] as const
 
 export type ReportReason = (typeof REPORT_REASONS)[number]
-
-// Seller reports use a subset: listing-specific reasons (duplicate,
-// wrong_category) make no sense for a person.
-export const SELLER_REPORT_REASONS: ReportReason[] = [
-  "spam",
-  "fraud",
-  "offensive_content",
-  "other",
-]
 
 export const REPORT_REASON_LABELS: Record<ReportReason, string> = {
   spam: "Spam",
@@ -66,6 +59,33 @@ export const REPORT_NOTE_MAX = 1000
 // show a friendly message before the request lands.
 export const REPORT_RATE_LIMIT_COUNT = 5
 export const REPORT_RATE_LIMIT_WINDOW_MS = 60 * 60 * 1000
+
+// Server-action input schema (shared with the action + unit tests, so the XOR
+// target rule is exercised without importing a "use server" module).
+export const submitReportSchema = z
+  .object({
+    listingId: z.string().uuid().optional(),
+    sellerId: z.string().uuid().optional(),
+    reason: z.enum(REPORT_REASONS),
+    note: z
+      .string()
+      .max(
+        REPORT_NOTE_MAX,
+        `Keep the note under ${REPORT_NOTE_MAX} characters`
+      )
+      .optional(),
+  })
+  .refine(
+    (data) => {
+      const hasListing = Boolean(data.listingId)
+      const hasSeller = Boolean(data.sellerId)
+      // Exactly one target — the DB reports_target_one constraint requires XOR,
+      // so a dual (or empty) target is rejected here with a field-level error
+      // instead of a generic SQL violation (audit P1.17, #84).
+      return hasListing !== hasSeller
+    },
+    { message: "A report must target a listing or a seller, but not both" }
+  )
 
 // The login redirect target for the Report button, shared across signed-out
 // CTAs (see lib/nav.ts for the single definition).
